BACKEND/server-sacra360/HTR-service/requirements.txt
# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Storage (MinIO)
minio==7.2.0

# Procesamiento de imágenes y OCR/HTR
# Versiones compatibles entre sí
opencv-python-headless==4.10.0.84
numpy==1.26.4
easyocr==1.7.2
pdf2image==1.17.0
Pillow==10.0.1
pandas==2.2.3
PyMuPDF==1.23.8
# torch y torchvision se instalan en Dockerfile con --index-url específico

# Utilidades
python-dotenv==1.0.0
requests==2.31.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

